feat(nixos): separate nixos and home-manager
I decided to use standalone home-manager instead of using it as a NixOS
module. Mainly because it will be easier to add `nix` into my non NixOs
machines. The drawback is that I will need to do `home-manager switch`
command in addition to just `sudo nixos-rebuild switch` but this also
has some benefits like not making new boot menu for user configuration
changes and having the ability to just rollback user configuration or
system configuration independently.