feat(home): `starship` is causing left side of the prompt acting weird
The default emoji of `nix_shell` Starship module is bugged when put
after `fill` module. Basically it shift the shell to the left and
causing the left side of the prompt showing `...` and cut off the first
three characters.

Not sure if this is in `fish` shell only as I don't use Starship in
other shell.

Ref: https://github.com/starship/starship/issues/5924

Signed-off-by: budimanjojo <budimanjojo@gmail.com>